requirements.txt
Django==3.1
django-crispy-forms==1.9.2
asgiref==3.2.10
docutils==0.15.2
jmespath==0.10.0
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
